Review fixes

+++ app/src/main/java/com/example/revoluttask/data/model/BasicCurrencyRate.kt
@@ -1,6 +1,8 @@
 package com.example.revoluttask.data.model
 
+import java.math.BigDecimal
+
 data class BasicCurrencyRate(
     val tickerString: String,
-    val rate: Double
+    val rate: BigDecimal
 )
